Add Infestation (L8 Necromancy) to the Necronomicon
Applies the "infested" enchantment to affected monsters (in a smite-targeted
radius 2 AoE), making them explode into death scarabs on death. The number of
scarabs created depends on monster HD, up to a max of 3 at 27 HD (and can be
zero for low-HD monsters).

Affects anything, including undead and demons - thematically, the death scarabs
can feed on whatever force is animating the monster (or something), since the
aim is for it to be a necromancy spell that still has use later on (but with
the downside that you still need some means of killing the targets in the first
place for it to trigger).